requirements.txt
Flask
PILLOW
matplotlib
numpy
pandas
requests
scikit-learn
tensorflow
openpyxl
opencv-python